Move dispatchDecidePolicyForNavigationAction to a location that can be shared by RemoteFrameClient https://bugs.webkit.org/show_bug.cgi?id=256918
Reviewed by Chris Dumez.

Just moving code around.  No change in behavior.

* Source/WebCore/Headers.cmake:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/loader/FrameLoaderClient.h: Copied from Source/WebKit/WebProcess/WebCoreSupport/WebRemoteFrameClient.h.
* Source/WebCore/loader/LocalFrameLoaderClient.h:
* Source/WebKit/Shared/WebHitTestResultData.cpp:
(WebKit::WebHitTestResultData::fromNavigationActionAndLocalFrame):
* Source/WebKit/Shared/WebHitTestResultData.h:
* Source/WebKit/Sources.txt:
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:
* Source/WebKit/WebProcess/WebCoreSupport/WebFrameLoaderClient.cpp: Added.
(WebKit::WebFrameLoaderClient::WebFrameLoaderClient):
(WebKit::WebFrameLoaderClient::dispatchDecidePolicyForNavigationAction):
* Source/WebKit/WebProcess/WebCoreSupport/WebFrameLoaderClient.h: Copied from Source/WebKit/WebProcess/WebCoreSupport/WebRemoteFrameClient.h.
(WebKit::WebFrameLoaderClient::webFrame const):
* Source/WebKit/WebProcess/WebCoreSupport/WebLocalFrameLoaderClient.cpp:
(WebKit::WebLocalFrameLoaderClient::WebLocalFrameLoaderClient):
(WebKit::WebLocalFrameLoaderClient::dispatchDecidePolicyForNewWindowAction):
(WebKit::WebLocalFrameLoaderClient::dispatchDecidePolicyForNavigationAction):
(WebKit::webHitTestResultDataInNavigationActionData): Deleted.
* Source/WebKit/WebProcess/WebCoreSupport/WebLocalFrameLoaderClient.h:
* Source/WebKit/WebProcess/WebCoreSupport/WebRemoteFrameClient.cpp:
(WebKit::WebRemoteFrameClient::WebRemoteFrameClient):
* Source/WebKit/WebProcess/WebCoreSupport/WebRemoteFrameClient.h:

Canonical link: https://commits.webkit.org/264183@main
