+++ Source/WebKit/Shared/WebHitTestResultData.h
@@ -46,6 +46,8 @@ class Encoder;
 
 namespace WebCore {
 class HitTestResult;
+class LocalFrame;
+class NavigationAction;
 }
 
 namespace WebKit {
@@ -108,6 +110,10 @@ struct WebHitTestResultData {
     WebCore::IntRect elementBoundingBoxInWindowCoordinates(const WebCore::HitTestResult&);
 
     std::optional<WebKit::SharedMemory::Handle> getImageSharedMemoryHandle() const;
+
+#if PLATFORM(MAC) || HAVE(UIKIT_WITH_MOUSE_SUPPORT)
+    static std::optional<WebHitTestResultData> fromNavigationActionAndLocalFrame(const WebCore::NavigationAction&, WebCore::LocalFrame*);
+#endif
 };
 
 } // namespace WebKit
