+++ Source/WebKit/WebProcess/WebCoreSupport/WebRemoteFrameClient.cpp
@@ -36,7 +36,7 @@
 namespace WebKit {
 
 WebRemoteFrameClient::WebRemoteFrameClient(Ref<WebFrame>&& frame, ScopeExit<Function<void()>>&& frameInvalidator)
-    : m_frame(WTFMove(frame))
+    : WebFrameLoaderClient(WTFMove(frame))
     , m_frameInvalidator(WTFMove(frameInvalidator))
 {
 }
